Fix player list loading

+++ src/game/Player.cpp
@@ -1572,7 +1572,7 @@ bool Player::BuildEnumData( QueryResult * result, WorldPacket * p_data )
     *p_data << fields[11].GetFloat();                       // y
     *p_data << fields[12].GetFloat();                       // z
 
-    *p_data << uint32(fields[13].GetUInt32());              // guild id
+    *p_data << uint64(fields[13].GetUInt32());              // guild id
 
     uint32 char_flags = 0;
     uint32 playerFlags = fields[14].GetUInt32();
